♻️ ref(issue-link): Move UISchema-to-Form logic to separate component
